Refactor para recibir duration en controller

+++ BetterCalm/Backend/MSP.BetterCalm.WebAPI/Dtos/ScheduleMeetingDto.cs
@@ -7,6 +7,6 @@ namespace MSP.BetterCalm.WebAPI.Dtos
         public Patient Patient { get; set; }
         public Problematic Problematic { get; set; }
         
-        public decimal Duration { get; set; }
+        public double Duration { get; set; }
     }
 }
